+++ NEWS.md
@@ -1,3 +1,21 @@
+# spmodel 0.2.0
+
+## Minor updates
+
+* `splm()` and `spautor()` allow multiple models to be fit when the `spcov_type` argument is a vector of length greater than one or the `spcov_initial` argument is a list of length greater than one.
+    * The resulting object is a list with class `spmod_list`. Each element of the list holds a different model fit.
+    * `glances()` is used on an `spmod_list` object to glance at each model fit.
+    * `predict()` is used on an `spmod_list` object to predict at the locations in `newdata` for each model fit.
+* Added `splmRF()` and `spautorRF()` functions to fit random forest residual spatial linear models that supplement random forest models by incorporating spatial covariance among the residuals.
+    * The resulting object has class `spmodRF` (one spatial covariance) or `spmodRF_list` (multiple spatial covariances)
+    * These objects are built for use with `predict()` to perform random forest residual Kriging (prediction).
+* Minor vignette updates.
+* Minor documentation updates.
+
+## Bug Fixes
+
+* Fixed a bug that prevents display of spatial covariance type in summary of `.spmod` objects.
+
 # spmodel 0.1.1
 
 ## Minor updates